requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.7
cryptography==41.0.4
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==2.3.7
